Extract handleBlockCollision and handleEntityCollision from Projectile implementations into AbstractProjectile superclass
